// src/appearance/enums/appearance_category_code.enum.ts
export enum AppearanceCategoryCode {
	attic = "attic",
	favorites = "favorites",
	underwear = "underwear",
	skin = "skin",
	tatoo = "tatoo",
	mouth = "mouth",
	eye = "eye",
	hair = "hair",
	sock = "sock",
	shoe = "shoe",
	pants = "pants",
	handAccessory = "handAccessory",
	top = "top",
	coat = "coat",
	glove = "glove",
	necklace = "necklace",
	dress = "dress",
	hat = "hat",
	faceAccessory = "faceAccessory",
	background = "background",
	belt = "belt",
	ambient = "ambient",
}